+++ package.json
@@ -19,6 +19,8 @@
     "classnames": "^2.5.1",
     "framer-motion": "^11.0.3",
     "js-cookie": "^3.0.5",
+    "lodash": "^4.17.21",
+    "lodash.debounce": "^4.0.8",
     "next": "14.0.4",
     "next-redux-wrapper": "^8.1.0",
     "node": "^20.10.0",
@@ -42,6 +44,7 @@
     "@babel/core": "^7.23.7",
     "@babel/preset-env": "^7.23.8",
     "@babel/preset-react": "^7.23.3",
+    "@types/lodash": "^4.14.202",
     "@types/node": "^20",
     "@types/react": "^18",
     "@types/react-dom": "^18",
@@ -50,6 +53,7 @@
     "babel-loader": "^9.1.3",
     "eslint": "^8.56.0",
     "eslint-config-next": "14.0.4",
+    "lodash-webpack-plugin": "^0.11.6",
     "postcss": "^8.4.33",
     "postcss-cli": "^11.0.0",
     "prettier": "^3.2.3",
